[!!!][TASK] Improve CLi command and type hinting
This adds an artificial "lifetime" to the
event processing CLI command in order to not run
into typical long-running-PHP-processes problems
like timing out DB connections and so on.

You will need to daemonize the respective command
in order to respawn it whenever it quits intentionally.

This would be possible via ``daemon``, ``upstart`` or
``supervisord``.

This is considered breaking because you now need
to take care to respawn that process.