[!!!][TASK] Move actual event handling to sub process
This introduces a new command ``ag.event:event.handle``
which will actually handle a given event with a given
event handler. The initial queue process will internally
call this command as a sub process.

This is necessary in order to avoid the problems the
long running PHP process will lead to, like, for example,
Singletons staying the same all the time, the Entity Manager
with its Identity Map not fetching the current peristence
state; etc.

This is marked breaking in cases where the sub process
invocation might need some additional configuration to work
on the system it's running on.

Resolves #11